# Implement synthetic benchmark