--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -1,10 +1,10 @@
 name: CI
 on:
   push:
-    branches: ['main']
-    tags: ['*']
+    branches: ["main"]
+    tags: ["*"]
   pull_request:
-    branches: ['main']
+    branches: ["main"]
 
 jobs:
   test:
@@ -18,6 +18,7 @@ jobs:
       - run: npm run build
       - run: npm run types -- --noEmit
       - run: npm test
+      - run: npm run benchmark
 
   publish:
     name: Publish