--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,4 @@
 node_modules
 *.js
 *.d.ts
+!benchmarks/*.js