--- a/benchmarks/run.js
+++ b/benchmarks/run.js
@@ -21,6 +21,7 @@ for (let i = 0; i < 1000; i++) {
 
 let result = null;
 
+console.time('simple')
 performance.mark('start-simple')
 
 for (let i = 0; i < 10_000; i++) {
@@ -28,9 +29,11 @@ for (let i = 0; i < 10_000; i++) {
 }
 
 performance.mark('end-simple')
+console.timeEnd('simple')
+
 assert.deepEqual(result, [0])
 
-
+console.time('wildcard')
 performance.mark('start-wildcard')
 
 for (let i = 0; i < 10_000; i++) {
@@ -38,11 +41,14 @@ for (let i = 0; i < 10_000; i++) {
 }
 
 performance.mark('end-wildcard')
+console.timeEnd('wildcard')
+
 assert.equal(result[0], 'prop0.value')
 assert.equal(result[1], 'prop1.value')
 assert.equal(result[2], 'prop2.value')
 assert.equal(result[999], 'prop999.value')
 
+console.time('globstar')
 performance.mark('start-globstar')
 
 for (let i = 0; i < 100; i++) {
@@ -50,6 +56,7 @@ for (let i = 0; i < 100; i++) {
 }
 
 performance.mark('end-globstar')
+console.timeEnd('globstar')
 
 assert.ok(result.includes('prop0.value'))
 assert.ok(result.includes('prop0.nested.value'))
@@ -60,6 +67,8 @@ assert.ok(result.includes(`prop0${'.nested'.repeat(99)}.value`))
 assert.ok(result.includes('prop999.nested.value'))
 
 console.log(
+  '\n',
+  'Details:\n',
   performance.measure('Simple x10_000', 'start-simple', 'end-simple'),
   performance.measure('Wildcard x10_000', 'start-wildcard', 'end-wildcard'),
   performance.measure('Globstar x100', 'start-globstar', 'end-globstar'),