--- a/package.json
+++ b/package.json
@@ -44,6 +44,6 @@
     "typescript": "^4.9.5"
   },
   "dependencies": {
-    "glob-to-regexp": "^0.4.1"
+    "micromatch": "^4.0.5"
   }
 }