--- a/package.json
+++ b/package.json
@@ -45,6 +45,6 @@
     "typescript": "^4.9.5"
   },
   "dependencies": {
-    "micromatch": "^4.0.5"
+    "glob-to-regexp": "^0.4.1"
   }
 }