--- a/package.json
+++ b/package.json
@@ -12,7 +12,7 @@
   "scripts": {
     "test": "jest src/",
     "build": "swc src/ -d . --config-file .swcrc",
-    "types": "tsc --esModuleInterop --declaration --outDir . src/index.ts",
+    "types": "tsc --skipLibCheck --esModuleInterop --declaration --outDir . src/index.ts",
     "prepublishOnly": "npm run build && npm run types -- --emitDeclarationOnly",
     "prettier": "prettier --write \"**/*.{ts,json,md}\"",
     "lint": "prettier --check \"**/*.{ts,json,md}\"",
@@ -45,6 +45,6 @@
     "typescript": "^4.9.5"
   },
   "dependencies": {
-    "glob-to-regexp": "^0.4.1"
+    "minimatch": "^9.0.3"
   }
 }