--- a/src/glob.ts
+++ b/src/glob.ts
@@ -1,6 +1,4 @@
-import globToRegExp from "glob-to-regexp";
-
-// import { globToRegExp } from './deno-glob';
+import micromatch from "micromatch";
 
 export function globPaths(globPattern: string, obj: any): string[] {
   return glob(globPattern, obj, "path");
@@ -42,7 +40,7 @@ export function glob(globPattern: string, obj: any, mode: "path" | "value"): any
           );
           globByDepth.set(path.length, partialMatcher);
         }
-        const isPartialMatch = partialMatcher.test(currentPath.join("/"));
+        const isPartialMatch = partialMatcher(currentPath.join("/"));
 
         if (isPartialMatch) {
           traverse(value, currentPath);
@@ -56,14 +54,14 @@ export function glob(globPattern: string, obj: any, mode: "path" | "value"): any
   return result;
 }
 
-const objectPathMatches = (pathGlob: RegExp, paths: string[]) => {
+const objectPathMatches = (pathGlob: ReturnType<micromatch.matcher>, paths: string[]) => {
   const path = paths.join("/");
 
-  return pathGlob.test(path);
+  return pathGlob(path);
 };
 
 const toPathRegex = (glob: string) => {
   const pathGlob = glob.split(".").join("/"); // replace all dots with slashes
 
-  return globToRegExp(pathGlob, { extended: true, globstar: true });
+  return micromatch.matcher(pathGlob, { dot: true, star: true });
 };