… .swcrc