Baseline results running benchmark on GitHub Actions:

```
simple: 2.051s
wildcard: 11.318s
globstar: 1:01.684 (m:ss.mmm)
```

With micromatch, slightly worse:

```
simple: 2.807s
wildcard: 13.961s
globstar: 1:02.495 (m:ss.mmm)
```

With optimizations, skipping work and primitive match precheck:

```
simple: 2.630s
wildcard: 11.224s
globstar: 44.068s
```